Fixes Issue 45 with regards to reuse of database with multiple datasets.
This commit fixes what was tried to be fixed in Issue 44. In this case,
the situation is that Saga no longer allows for internet access on the
compute nodes. This means that we cannot download data within a slurm job.
Thus we have to preload the databases somewhere, and use those. This
commit sets up provisions for using a database that is preloaded on saga.